requirements.txt
pdfminer.six
spacy